Refactor by using fmu.tools.rms function copy_rms_param
